move KeySelectionValueListAttribute to manager
