+++ src/es6/canvas_widget/Manager.js
@@ -18,6 +18,8 @@ import "../lib/jquery/jquery.autoGrowInput";
 import { OpenAppProvider } from "../lib/openapp";
 import { default as ActivityOperation } from "../operations/non_ot/ActivityOperation";
 import {
+  AttributeAddOperation,
+  AttributeDeleteOperation,
   EdgeAddOperation,
   EdgeDeleteOperation,
   NodeAddOperation,
@@ -34,7 +36,6 @@ import { default as AbstractValue } from "./AbstractValue";
 import Arrows from "./Arrows";
 import { default as BooleanAttribute } from "./BooleanAttribute";
 import { default as FileAttribute } from "./FileAttribute";
-import { default as KeySelectionValueListAttribute } from "./KeySelectionValueListAttribute";
 import KeySelectionValueSelectionValueListAttribute from "./KeySelectionValueSelectionValueListAttribute";
 import QuizAttribute from "./QuizAttribute";
 import { default as SingleColorValueAttribute } from "./SingleColorValueAttribute";
@@ -49,6 +50,11 @@ import LogicalConjunctions from "./viewpoint/LogicalConjunctions";
 import LogicalOperator from "./viewpoint/LogicalOperator";
 import RenamingListAttribute from "./viewpoint/RenamingListAttribute";
 import ViewTypesUtil from "./viewpoint/ViewTypesUtil";
+
+const keySelectionValueListAttributeHtml = await loadHTML(
+  "../../templates/canvas_widget/list_attribute.html",
+  import.meta.url
+);
 
 const keySelectionValueAttributeHtml = await loadHTML(
   "../../templates/canvas_widget/key_value_attribute.html",
@@ -8347,6 +8353,324 @@ export class KeySelectionValueAttribute extends AbstractAttribute {
 
     _$node.find(".key").append(_key.get$node());
     _$node.find(".value").append(_value.get$node());
+  }
+}
+
+/**
+ * KeySelectionValueListAttribute
+ * @class canvas_widget.KeySelectionValueListAttribute
+ * @extends canvas_widget.AbstractAttribute
+ * @memberof canvas_widget
+ * @constructor
+ * @param {string} id Entity id
+ * @param {string} name Name of attribute
+ * @param {AbstractEntity} subjectEntity Entity the attribute is assigned to
+ * @param {Object} options Selection options
+ */
+export class KeySelectionValueListAttribute extends AbstractAttribute {
+  static TYPE = "KeySelectionValueListAttribute";
+  constructor(id, name, subjectEntity, options) {
+    super(id, name, subjectEntity);
+    var that = this;
+
+    /**
+     * Selection options
+     * @type {Object}
+     * @private
+     */
+    var _options = options;
+
+    /**
+     * List of attributes
+     * @type {Object}
+     * @private
+     */
+    var _list = {};
+
+    /**
+     * jQuery object of DOM node representing the attribute
+     * @type {jQuery}
+     * @private
+     */
+    var _$node = $(_.template(keySelectionValueListAttributeHtml)());
+
+    /**
+     * Inter widget communication wrapper
+     * @type {Object}
+     */
+    y = y || window.y;
+    var _iwcw = IWCW.getInstance(CONFIG.WIDGET.NAME.MAIN, y);
+
+    /**
+     * Apply an Attribute Add Operation
+     * @param {operations.ot.AttributeAddOperation} operation
+     */
+    var processAttributeAddOperation = function (operation) {
+      var attribute = new KeySelectionValueAttribute(
+        operation.getEntityId(),
+        "Attribute",
+        that,
+        _options
+      );
+      attribute.registerYMap();
+      that.addAttribute(attribute);
+      if (_$node.find(".list").find("#" + attribute.getEntityId()).length == 0)
+        _$node.find(".list").append(attribute.get$node());
+    };
+
+    /**
+     * Apply an Attribute Delete Operation
+     * @param {operations.ot.AttributeDeleteOperation} operation
+     */
+    var processAttributeDeleteOperation = function (operation) {
+      var attribute = that.getAttribute(operation.getEntityId());
+      if (attribute) {
+        that.deleteAttribute(attribute.getEntityId());
+        attribute.get$node().remove();
+      }
+    };
+
+    /**
+     * Propagate an Attribute Add Operation to the remote users and the local widgets
+     * @param {operations.ot.AttributeAddOperation} operation
+     */
+    var propagateAttributeAddOperation = function (operation) {
+      processAttributeAddOperation(operation);
+    };
+
+    /**
+     * Propagate an Attribute Delete Operation to the remote users and the local widgets
+     * @param {operations.ot.AttributeDeleteOperation} operation
+     */
+    var propagateAttributeDeleteOperation = function (operation) {
+      processAttributeDeleteOperation(operation);
+      var ymap = that.getRootSubjectEntity().getYMap();
+      ymap.delete(operation.getEntityId() + "[key]");
+    };
+
+    /**
+     * Callback for a remote Attrbute Add Operation
+     * @param {operations.ot.AttributeAddOperation} operation
+     */
+    var remoteAttributeAddCallback = function (operation) {
+      if (
+        operation instanceof AttributeAddOperation &&
+        operation.getRootSubjectEntityId() ===
+          that.getRootSubjectEntity().getEntityId() &&
+        operation.getSubjectEntityId() === that.getEntityId()
+      ) {
+        processAttributeAddOperation(operation);
+      }
+    };
+
+    /**
+     * Callback for a remote Attribute Delete Operation
+     * @param {operations.ot.AttributeDeleteOperation} operation
+     */
+    var remoteAttributeDeleteCallback = function (operation) {
+      if (
+        operation instanceof AttributeDeleteOperation &&
+        operation.getRootSubjectEntityId() ===
+          that.getRootSubjectEntity().getEntityId() &&
+        operation.getSubjectEntityId() === that.getEntityId()
+      ) {
+        processAttributeDeleteOperation(operation);
+      }
+    };
+
+    var localAttributeAddCallback = function (operation) {
+      if (
+        operation instanceof AttributeAddOperation &&
+        operation.getRootSubjectEntityId() ===
+          that.getRootSubjectEntity().getEntityId() &&
+        operation.getSubjectEntityId() === that.getEntityId()
+      ) {
+        propagateAttributeAddOperation(operation);
+      }
+    };
+    /**
+     * Callback for a local Attribute Delete Operation
+     * @param {operations.ot.AttributeDeleteOperation} operation
+     */
+    var localAttributeDeleteCallback = function (operation) {
+      if (
+        operation instanceof AttributeDeleteOperation &&
+        operation.getRootSubjectEntityId() ===
+          that.getRootSubjectEntity().getEntityId() &&
+        operation.getSubjectEntityId() === that.getEntityId()
+      ) {
+        propagateAttributeDeleteOperation(operation);
+      }
+    };
+
+    /**
+     * Add attribute to attribute list
+     * @param {canvas_widget.AbstractAttribute} attribute
+     */
+    this.addAttribute = function (attribute) {
+      var id = attribute.getEntityId();
+      if (!_list.hasOwnProperty(id)) {
+        _list[id] = attribute;
+      }
+    };
+
+    /**
+     * Get attribute of attribute list by its entity id
+     * @param id
+     * @returns {canvas_widget.AbstractAttribute}
+     */
+    this.getAttribute = function (id) {
+      if (_list.hasOwnProperty(id)) {
+        return _list[id];
+      }
+      return null;
+    };
+
+    /**
+     * Delete attribute from attribute list by its entity id
+     * @param {string} id
+     */
+    this.deleteAttribute = function (id) {
+      if (_list.hasOwnProperty(id)) {
+        delete _list[id];
+      }
+    };
+
+    /**
+     * Get attribute list
+     * @returns {Object}
+     */
+    this.getAttributes = function () {
+      return _list;
+    };
+
+    /**
+     * Set attribute list
+     * @param {Object} list
+     */
+    this.setAttributes = function (list) {
+      _list = list;
+    };
+
+    /**
+     * Get jQuery object of the DOM node representing the attribute (list)
+     * @returns {jQuery}
+     */
+    this.get$node = function () {
+      return _$node;
+    };
+
+    /**
+     * Get JSON representation of the attribute (list)
+     * @returns {Object}
+     */
+    this.toJSON = function () {
+      var json = AbstractAttribute.prototype.toJSON.call(this);
+      json.type = KeySelectionValueListAttribute.TYPE;
+      var attr = {};
+      _.forEach(this.getAttributes(), function (val, key) {
+        attr[key] = val.toJSON();
+      });
+      json.list = attr;
+      return json;
+    };
+
+    /**
+     * Set attribute list by its JSON representation
+     * @param json
+     */
+    this.setValueFromJSON = function (json) {
+      _.forEach(json.list, function (val, key) {
+        var attribute = new KeySelectionValueAttribute(
+          key,
+          key,
+          that,
+          _options
+        );
+        attribute.setValueFromJSON(json.list[key]);
+        that.addAttribute(attribute);
+        if (
+          _$node.find(".list").find("#" + attribute.getEntityId()).length == 0
+        )
+          _$node.find(".list").append(attribute.get$node());
+      });
+    };
+
+    /**
+     * Register inter widget communication callbacks
+     */
+    this.registerCallbacks = function () {
+      _iwcw.registerOnDataReceivedCallback(localAttributeAddCallback);
+      _iwcw.registerOnDataReceivedCallback(localAttributeDeleteCallback);
+    };
+
+    /**
+     * Unregister inter widget communication callbacks
+     */
+    this.unregisterCallbacks = function () {
+      _iwcw.unregisterOnDataReceivedCallback(localAttributeAddCallback);
+      _iwcw.unregisterOnDataReceivedCallback(localAttributeDeleteCallback);
+    };
+
+    _$node.find(".name").text(this.getName());
+
+    for (var attributeId in _list) {
+      if (_list.hasOwnProperty(attributeId)) {
+        _$node.find(".list").append(_list[attributeId].get$node());
+      }
+    }
+
+    if (_iwcw) {
+      that.registerCallbacks();
+    }
+
+    this.registerYMap = function () {
+      var ymap = that.getRootSubjectEntity().getYMap();
+      var attrs = that.getAttributes();
+      for (var key in attrs) {
+        if (attrs.hasOwnProperty(key)) {
+          var attr = attrs[key];
+          attr.getKey().registerYType();
+          attr.getValue().registerYType();
+        }
+      }
+
+      ymap.observe(function (event) {
+        const changedKeys = event.keysChanged;
+        const array = Array.from(event.changes.keys.entries());
+        array.forEach(([key, change]) => {
+          if (key.indexOf("[key]") != -1) {
+            var operation;
+            var data = event.currentTarget.get(key);
+            switch (change.action) {
+              case "add": {
+                //  var yUserId = event.object.map[key][0];
+                //  if (yUserId === y.clientID) return;
+                operation = new AttributeAddOperation(
+                  key.replace(/\[\w*\]/g, ""),
+                  that.getEntityId(),
+                  that.getRootSubjectEntity().getEntityId(),
+                  that.constructor.name
+                );
+                remoteAttributeAddCallback(operation);
+
+                break;
+              }
+              case "delete": {
+                operation = new AttributeDeleteOperation(
+                  key.replace(/\[\w*\]/g, ""),
+                  that.getEntityId(),
+                  that.getRootSubjectEntity().getEntityId(),
+                  that.constructor.name
+                );
+                remoteAttributeDeleteCallback(operation);
+                break;
+              }
+            }
+          }
+        });
+      });
+    };
   }
 }
